pyproject.toml
[tool.poetry]
name = "most-used-langchain-chains"
version = "0.1.0"
description = ""
authors = ["full-stack-ai <fullstackinfo.ai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.125.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
beautifulsoup4 = "^4.12.3"
langchain-community = "^0.0.31"
langchain-openai = "^0.1.1"
python-dotenv = "^1.0.1"
langchain = "^0.1.14"
fastapi = "^0.110.1"
httpx = "^0.27.0"
uvicorn = "^0.29.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
